refactor: component 리팩토링
- AuthComponent와 UserComponent 메서드 위치 재조정
